+++ CHANGELOG.md
@@ -15,6 +15,13 @@ Types of changes:
 - `Security` in case of vulnerabilities.
 -->
 
+## [Unreleased]
+
+### Added
+
+- Standard [interfaces](https://docs.julialang.org/en/v1/manual/interfaces/):
+  - iteration,
+
 ## [0.1.0]
 
 ### Added
@@ -31,5 +38,6 @@ Types of changes:
 
 Empty project.
 
+[Unreleased]: https://github.com/FedericoStra/ProtectedArrays.jl/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/FedericoStra/ProtectedArrays.jl/compare/v0.0.0...v0.1.0
 [0.0.0]: https://github.com/FedericoStra/ProtectedArrays.jl/releases/tag/v0.0.0